Error Fixes: Directory paths and LIST
Hi, I made 2 fixes but I wanted your team's input on fix 2.

1) For LIST, the returned IDs should be a list of int not strings. I fixed this by replacing std::vector<string> with std::vector<int>.

2) API request handler wasn't working when testing with the actual disk. The issue was fs::path dir in disk_file_store.cc was trying to read/write to absolute path /mnt/crud, which caused permission denied errors.

I changed it to relative path ../mnt/crud to resolve the bug. However this does assume you run in one of the root repository's subdirectories like /build or /tests if you want /mnt/crud to be placed in the root repository.

Change-Id: I426d5fdf55d6fc31311df08dd1eb25bade317592